SCP-1921: MockNode: tidying up for concistency
- Moving all types to Types module
- organizing imports (grouped external/internal)
